Fix player revival system and enhance test infrastructure
Key fixes and improvements:
- Fixed start room initialization: changed from non-existent 1000 to existing 3001 (Forest Temple)
- Verified comprehensive dead player command whitelist is working correctly
- Confirmed prompt system fully functional with health/movement/combat status
- Reorganized test infrastructure with clear unit/integration categories
- Updated test documentation to reflect current system state
- Fixed container system segfaults (some test failures remain to be addressed)

All core player lifecycle systems now working correctly.
Next iteration should focus on container system completion.

+++ tests/test_harness.hpp
@@ -73,7 +73,7 @@ class TestHarness {
 
       private:
         TestHarness &harness_;
-        static inline int next_entity_id = 100;
+        static inline uint64_t next_entity_id = 100UL;
     };
 
     asio::io_context io_context;
